Add pollTimeout to Server
Server uses log polling for getting new request from a message broker.
pollTimeout sets the limit for waiting for new messages.
Be careful with this setting, setting it to a high value would
block the Shutdown. The rule of thumb is to keep pollTimeout
small enough for a faster Shutdown and large enough to not flood your
message broker with a large number of requests.